Extract class to file: CaptureCommandOutput
